requirements.txt
kafka-python>=2.2.15
numpy>=1.21.2
aiokafka>=0.12.0